fix routing on stream tabs
