+++ website/src/routes/streams/[id]/+layout.ts
@@ -1,7 +1,7 @@
 export const ssr = false;
 
 import type { VideoStream } from '$components/streams/types';
-import type { PageLoad } from '../$types';
+import type { LayoutLoad } from './$types';
 
 export const load = (async ({ params, fetch, depends }) => {
     depends('stream:data');
@@ -10,6 +10,8 @@ export const load = (async ({ params, fetch, depends }) => {
 
     const fetchStream = async (): Promise<VideoStream> => {
         const response = await fetch(`/api/v1/video-streams/${streamId}`);
+
+        console.log('response', response);
         if (!response.ok) throw new Error(`Failed to fetch stream: ${response.statusText}`);
         return response.json();
     };
@@ -17,4 +19,4 @@ export const load = (async ({ params, fetch, depends }) => {
     return {
         stream: fetchStream(),
     };
-}) satisfies PageLoad;
+}) satisfies LayoutLoad;
