+++ website/src/routes/streams/[id]/+page.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export function load({ params }) {
+    throw redirect(307, `${params.id}/overview`);
+}
