+++ website/src/routes/streams/[id]/assets/+page.svelte
@@ -1,0 +1,2 @@
+<!-- This page probably won't exist. Default route should be elsewhere -->
+<p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p>
