+++ website/src/routes/streams/[id]/events/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+    import EventsTab from '$components/streams/events/events-tab.svelte';
+</script>
+
+<EventsTab />
