+++ website/src/routes/streams/[id]/overview/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+    import OverviewTab from '$components/streams/overview/overview-tab.svelte';
+</script>
+
+<OverviewTab />
